ui layout updates
