+++ md_editor/components/NewEmptyDoc.js
@@ -1,0 +1,15 @@
+export default function NewEmptyDoc() {
+  return (
+    <div>
+      <h1>Start Typing below!</h1>
+
+      <style jsx>{`
+        h1 {
+          font-size: 17px;
+          color: #a4bbbb;
+          text-align: center;
+        }
+      `}</style>
+    </div>
+  );
+}
